Change db owner type

+++ TelegramBot/Migrations/AppDbContextModelSnapshot.cs
@@ -26,7 +26,7 @@ namespace TelegramBot.Migrations
                         .IsRequired()
                         .HasColumnType("TEXT");
 
-                    b.Property<long>("OwnerId")
+                    b.Property<int>("OwnerId")
                         .HasColumnType("INTEGER");
 
                     b.Property<string>("Quote")
